CHROMIUM: platform: x86: add ACPI driver for ChromeOS.
This driver attaches to the Chromeos ACPI device and the exports the values
reported by the ACPI in a sysfs directory. All ACPI values are presented in
the string form (numbers as decimal values) and can be accessed as the
contents of the appropriate read only files in the sysfs directory tree
originating in /sys/devices/platform/chromeos_acpi.

[rebase54(gwendal): Use ACPI_NAMESEG_SIZE instead of ACPI_NAME_SIZE]
[rebase510(gwendal): Fix x86/Makefile, ---help in Kconfig]

Signed-off-by: Olof Johansson <olof@lixom.net>
(Squashed related commits bc14076e8807 .. 6addb52fb680)
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Change-Id: If201cae15f7005e3c2c440bc04f2622bf6b1daa9