FROMLIST: sched: Return idle task in pick_next_task for offlined CPU
There could be races in core scheduler where a CPU is trying to pick
a task for its sibling in core scheduler, when that CPU has just been
offlined.  We should not schedule any tasks on the CPU in this case.
Return an idle task in pick_next_task for this situation.

(am from https://lore.kernel.org/lkml/cover.1572437285.git.vpillai@digitalocean.com/T/)

BUG=b:152605392
TEST=run power_VideoCall test

Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Iff2d1b4c9dc2f4c882fd1c3f1a1b4fde7793ae52