FROMLIST: sched/core: Update core scheduler queue when taking cpu online/offline
When we bring a CPU online and enable core scheduler, tasks that need
core scheduling need to be placed in the core's core scheduling queue.
Likewise when we taks a CPU offline or disable core scheudling on a
core, tasks in the core's core scheduling queue need to be removed.
Without such mechanisms, the core scheduler causes OOPs due to
inconsistent core scheduling state of a task.

Implement such enqueue and dequeue mechanisms according to a CPU's change
in core scheduling status.  The switch of core scheduling mode of a core,
and enqueue/dequeue of tasks on a core's queue due to the core scheduling
mode change has to be run in a separate context as it cannot be done in
the context taking cpu online/offline.

BUG=b:152605392
TEST=run power_VideoCall test

(am from https://lore.kernel.org/lkml/cover.1572437285.git.vpillai@digitalocean.com/T/)

Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: If7422ad2d2677a03cca3cb4a0c9dd0812c9dc8e6