FROMLIST: sched: Add a per-thread core scheduling interface
Add a per-thread core scheduling interface which allows a thread to tag
itself and enable core scheduling. Based on discussion at OSPM with
maintainers, we propose a prctl(2) interface accepting values of 0 or 1.
 1 - enable core scheduling for the task.
 0 - disable core scheduling for the task.

Special cases:
(1)
The core-scheduling patchset contains a CGroup interface as well. In
order for us to respect users of that interface, we avoid overriding the
tag if a task was CGroup-tagged because the task becomes inconsistent
with the CGroup tag. Instead return -EBUSY.

(2)
If a task is prctl-tagged, allow the CGroup interface to override
the task's tag.

ChromeOS will use core-scheduling to securely enable hyperthreading.
This cuts down the keypress latency in Google docs from 150ms to 50ms
while improving the camera streaming frame rate by ~3%.

(am from https://lore.kernel.org/patchwork/patch/1245234/)

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
hange-Id: Icb565b6f0edf8b06fb7533c06bf0de7e1709b090
Change-Id: I037216ca3d45644caf0b0cd2efcc523cc29cfdbe