sched: Add a per-thread core scheduling interface
Add a per-thread core scheduling interface which allows a task to share
a core with another task.

Note that if CGroup interface is used for a task, the prctl(2) will
override it, and vice-versa. The user is expected to use only one of
these interfaces at a time.

ChromeOS will use core-scheduling to securely enable hyperthreading.
This cuts down the keypress latency in Google docs from 150ms to 50ms
while improving the camera streaming frame rate by ~3%.

XXX : sched_core_put() needs to be called if prctl(2) task is added to
tagged cgroup.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>