CHROMIUM: fs/namespace.c: WARN for old nosymfollow option
The nosymfollow mount option used to be implemented in the chromiumos
LSM.  Since it wasn't a recognized flag, it was passed in the data
parameter to mount() and parsed out by the LSM.

Commit ee330608ec71 ("FROMLIST: Add a "nosymfollow" mount option.")
added a MNT_NOSYMFOLLOW flag intended to replace this.
Commit 90a3c9336ad5 ("CHROMIUM: fs/namespace.c: Copy nosymfollow to
flags") then added a compatibility check to set MNT_NOSYMFOLLOW if the
older option was passed.

We want to start logging a warning when callers pass the old flag so
that they can be found and updated.  Some callers will pass the flag
both ways for cross-kernel compatibility, so only check/log the warning
if the new flag is not present.

BUG=b:152074038
TEST=tast run security.StatefulPartitionHardening

Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>
Change-Id: I8d2ca1059f99ec728e33ca20c4e6b7fbf6bb5561
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2207443
Reviewed-by: Ross Zwisler <zwisler@google.com>