CHROMIUM: sched: Add a latence_sensitive attribute to tasks
This is the sched counterpart of the proc patch. The proc interface's
value is used only if CONFIG_UCLAMP_TASK is enabled. The proc value
overrides the CGroup value, similar to how the clamp attributes work.

Upstream is moving to use uclamp which we would like. However, only 2 of
the 3 attributes are available in it (the clamp ones). The third
attribute - uclamp latency sensitive attribute is out-of-tree in Android
and available only for CPU controller CGroups. This patch tries to add
support for the third so that the 2 upstream ones can be used without
requiring CPU controller CGroups.

In ChromeOS, previously the latency sensitive attribute has been set
using schedtune CGroup. However the upstream uclamp interface uses CPU
controller CGroups which contains attributes we want (clamps) but does
not contain the latency sensitive attribute. Even if we carry the
out-of-tree Android patch to add the latency sensitive attribute to
uclamp-cgroups, we still cannot use it for our usecase. The reason is
currently, several threads in the 'foreground' CPU controller CGroup
don't have the schedtune prefer-idle (aka latency sensitive) attribute
set. If we set this attribute for 'foreground', then all foreground
tasks will get the attribute set, which we don't want since that could
cause excessive power drain.

The solution is to add a temporary "latency sensitive" attribute to
procfs until upstream figures out what this interface should look like.
These patches aim to make the code self contained in their own
translation units so as to make the maintenance burden minimal while
keeping products working.

BUG=147773166
TEST=boot

Change-Id: Ia54b678a6843502f77819b01e0338fdd60a0e3db
Signed-off-by: Joel Fernandes <joelaf@google.com>