FIXUP: CHROMIUM: sched: Add a per-thread core scheduling interface
There is a limitation that a running task doing the core scheduling
prctl(2) has to sleep at least once before core-scheduling is activated
on it. This is usually not a problem because the task will eventually
block on something or sleep. But it can be a problem if the task never
goes to sleep. Fix this by requeuing the task into the core scheduling
queue.

Also do some small clean ups, and make resetting of a tag as a
privileged operation to prevent malicious processes from untagging
themselves in the future. This is currently not possible with sandboxing
of prctl() but in the future we need to becareful with unsandboxed
processes being able to do so.

BUG=b:152605392
TEST=run power_VideoCall test

Reviewed-by: Joel Fernandes <joelaf@google.com>
Change-Id: I2f5b8e8405d39006a9e3092a47cab13442b639a9