sched: Fix CPU hotplug causing crashes in task selection logic
The selection logic does not run correctly if the current CPU is not in the
cpu_smt_mask (which it is not because the CPU is offlined when the stopper
finishes running and needs to switch to idle).  There are also other issues
fixed by the patch I think such as: if some other sibling set core_pick to
something, however the selection logic on current cpu resets it before
selecting. In this case, we need to run the task selection logic again to
make sure it picks something if there is something to run. We might end up
picking a different task but that's Ok.

Yet another issue was, if the stopper thread is an unconstrained pick,
then rq->core_pick is set. The next time task selection logic runs when
stopper needs to switch to idle, the current CPU is not in the smt_mask.
This causes the previous ->core_pick to be picked again which happens to
be the unconstrained task! so the stopper keeps getting selected
forever.

That and there are a few more safe guards and checks around
checking/setting rq->core_pick. To test it, I ran rcutorture and made it
tag all torture threads. Then ran it in hotplug mode (hotplugging every
200ms) and it hit the issue. Now it runs for an hour or so without
issue. (Torture testing debug changes: https://bit.ly/38htfqK ).

Various fixes were tried causing varying degrees of crashes.  Finally I found
that it is easiest to just add current CPU to the smt_mask's copy always.
This is so that task selection logic always runs on the current CPU which
called schedule().

Link: lore.kernel.org/r/20200414133559.GT20730@hirez.programming.kicks-ass.net
Reported-by: Tim Chen <tim.c.chen@linux.intel.com>
Reported-by: Vineeth Pillai <vpillai@digitalocean.com>
Reviewed-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>