CHROMIUM: config: Disable efi backend for pstore
Commit bf67fad19e49 ("efi: Use more granular check for availability for
variable services") and follow-ups changed the condition for when the
efi pstore backend is available. As a result, it becomes enabled on
chromeos systems, which prevents the 'ramoops' backend from initialising
because there can only be one.

Remove support for the efi pstore backend since it should not be used
anywhere.

BUG=b:177474908
TEST=`tast run <DUT> crash.Ephemeral.pre_oobe_collection` on grunt with
     5.10 kernel

Change-Id: If42a5d3d145b1b2873398e8d67ea9cb1c79dc242
Signed-off-by: Anand K Mistry <amistry@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2635402
Reviewed-by: Guenter Roeck <groeck@chromium.org>