FIXUP: FROMGIT: PM / devfreq: Add cpu based scaling support to passive governor
For some architectures, cpu_data is defined as macro. This results in
errors such as

include/linux/devfreq.h:331:27: note: in expansion of macro 'cpu_data'
     331 |  struct devfreq_cpu_data *cpu_data[NR_CPUS];
         |                           ^~~~~~~~
   In file included from include/linux/devfreq_cooling.h:13,
                    from drivers/devfreq/devfreq.c:14:
   include/linux/devfreq.h:332:1: warning: no semicolon at end of struct or union

The problem was already reported upstream back in March, which may be the
reason why the patch never made it into the upstream kernel.

Fix the problem by renaming the offending variable to cpudata.

The original patch also introduces a memory leak in the error path, as already
reported upstream. No attempt is made to fix that problem.

BUG=b:129039370
TEST=boot krane

Change-Id: I2d54496a874399dcca1ab6301d6797406bcde224
Disallow-Recycled-Builds: test-failures
Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Guenter Roeck <groeck@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2917193
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>