FROMGIT: drm/i915: apply WaEnableVGAAccessThroughIOPort as needed
Currently as the workaround is applied the screen flickers. As a result
we do not achieve seamless boot experience.

Avoiding the issue in the common use-case might be hard, although we can
resolve it for dual GPU setups - when the "other" GPU is primary and/or
outputs are connected to it.

With this I was able to get seamless experience on my Intel/Nvidia box,
running systemd-boot and sddm/Xorg. Note that the i915 driver is within
initrd while the Nvidia one is not.

Without this patch, the splash presented by systemd-boot (UEFI BGRT) is
torn down as the code-path kicks in, leaving the monitor blank until the
login manager starts.

Same issue were reported with plymouth/grub, although personally I
wasn't able to get them to behave on my setup.

Suggested-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210604154905.660142-1-emil.l.velikov@gmail.com
(cherry picked from commit a3af0140663dc335e5c18277a61bb99bfcb60694
 git://anongit.freedesktop.org/drm-intel drm-intel-next)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:182519720
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer. Build tested on others

Change-Id: Iefb4c980a07cfd82ee8847ec748fb99be7dbd0c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2988577
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>