FROMGIT: iommu/io-pgtable: Introduce unmap_pages() as a page table op
The io-pgtable code expects to operate on a single block or
granule of memory that is supported by the IOMMU hardware when
unmapping memory.

This means that when a large buffer that consists of multiple
such blocks is unmapped, the io-pgtable code will walk the page
tables to the correct level to unmap each block, even for blocks
that are virtually contiguous and at the same level, which can
incur an overhead in performance.

Introduce the unmap_pages() page table op to express to the
io-pgtable code that it should unmap a number of blocks of
the same size, instead of a single block. Doing so allows
multiple blocks to be unmapped in one call to the io-pgtable
code, reducing the number of page table walks, and indirect
calls.

Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Suggested-by: Will Deacon <will@kernel.org>
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Georgi Djakov <quic_c_gdjako@quicinc.com>
Link: https://lore.kernel.org/r/1623850736-389584-2-git-send-email-quic_c_gdjako@quicinc.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 374c15594c4ee0dfcceb38852bd43be09070f402
 https://git.kernel.org/pub/scm/linux/kernel/git/joro/iommu.git core)

BUG=b:188373249
TEST=mmc tests

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I4cae003518cce94ccb12e50991866b7d8e9ffce3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2956626
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
(cherry picked from commit 2f60744ca0a83d08b164a440e45b5ae9785d6dcc)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3082883
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>