FROMGIT: HID: input: do not report stylus battery state as "full"
The power supply states of discharging, charging, full, etc, represent
state of charging, not the capacity level of the battery (for which
we have a separate property). Current HID usage tables to not allow
for expressing charging state of the batteries found in generic
styli, so we should simply assume that the battery is discharging
even if current capacity is at 100% when battery strength reporting
is done via HID interface. In fact, we were doing just that before
commit 581c4484769e.

This change helps UIs to not mis-represent fully charged batteries in
styli as being charging/topping-off.

Fixes: 581c4484769e ("HID: input: map digitizer battery usage")
Reported-by: Kenneth Albanowski <kenalba@google.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit f4abaa9eebde334045ed6ac4e564d050f1df3013
 https://git.kernel.org/pub/scm/linux/kernel/git/hid/hid.git for-5.15/core)

BUG=b:192408877
TEST=Build

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I3f93a7ca8f8c0289ae67ad9e571cf4010fe412dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3033322
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Tested-by: Kenneth Albanowski <kenalba@google.com>