CHROMIUM: bluetooth: Track usermode connections
Track usermode connections so that we can notify drivers when new ACL
and SCO connections are made. This is necessary for drivers like btusb
which need to switch altsettings when sco connections are active.

This patch is not suitable for upstreaming because peeking at the HCI
stream for usermode is not a good design. Eventually, connection
notifications need to come from userspace via ioctls on the socket
interface itself.

BUG=b:221497057
TEST=HFP working with Floss

Change-Id: I208b7c21974d202690228ff23bfe9ab917fec8d2
Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3514513
Reviewed-by: Archie Pusaka <apusaka@chromium.org>