CHROMIUM: gpu: mali: Reorder GPU clock enumeration
To migrate to an uniformed init/devfreq behavior while taking the HW
difference into account, one workaround for now is to reorder the clock
enumeration so that we can easily determine whether a platform has the
optional PLL clock by counting the number of the available GPU clocks.

This assumption may break against the new HW design, however this
applies to all the existing MTK platforms as of today.

BUG=b:227544156
TEST=emerge Corsola 5.15 kernel and boot to login screen

Change-Id: Iae345c0b8915d8c00030a4a3f87c4d78ed77ed72
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3590494
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>