cred: Move call_rcu() to call_rcu_lazy()
This is required to prevent callbacks triggering RCU machinery too
quickly and too often, which adds more power to the system.

The put_cred_rcu() does some debugging that would be less effective with
a lengthy RCU delay. This should be OK assuming that significant testing
happens on CONFIG_RCU_LAZY=n kernels.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>