sched: Split scheduler execution context
Lets define the scheduling context as all the scheduler state in
task_struct and the execution context as all state required to run the
task.

Currently both are intertwined in task_struct. We want to logically
split these such that we can run the execution context of one task
with the scheduling context of another.

To this purpose introduce rq::proxy to point to the task_struct used
for scheduler state and preserve rq::curr to denote the execution
context.

XXX connoro: A couple cases here may need more discussion:
- sched_yield() and yield_to(): whether we call the sched_class
  methods for rq->curr or rq->proxy, there seem to be cases where
  proxy exec could cause a yielding mutex owner to run again
  immediately. How much of a concern is this?
- push_rt_task() calls find_lowest_rq() which can only be invoked on
  RT tasks (scheduler context) but considers CPU affinity (execution
  context). For now we call find_lowest_rq() on rq->curr, bailing out
  when it is not RT, but does this give the desired behavior?

[added lot of comments/questions - identifiable by XXX]
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lkml.kernel.org/r/20181009092434.26221-5-juri.lelli@redhat.com
[add additional comments and update more sched_class code to use
 rq::proxy]
Signed-off-by: Connor O'Brien <connoro@google.com>